refactor: moves dnd logic to 'useDragAndDrop' hook
